copypi/requirements.txt
epd2in13>=0.1
paho-mqtt>=1.1
Pillow>=4.3.0
